Disable poking at other Thumb insns, for FuzzBALL
This code breaks the abstraction that VEX looks only at the code
that you pass to it. That's not a problem in Valgrind because the
program exists in the "host" at its "guest" address, but FuzzBALL
has a bigger guest/host separation, so trying to look at other instructions
just won't work (typically it crashes).