Disable special case for call-pop idiom, for FuzzBALL
Disabled by treating it as a special case of jump-following, which
FuzzBALL already disables. This special case is not a safe optimization
if other code looks at your stack. But the more proximate problem for
us was that it was triggering even when we only wanted to translate
the call on its own.